Fix,Refactor: app.js POST 작업 모듈화 하기
